Fix special characters creating invalid file name
Valid Windows file names cannot contain | ? : * " < > \  characters.

Slugify does not remove these and the most obvious would be the '?' in a
query string. When these characters are present in the URL provided, the
file fails to get written to disk without any warning or error.

Amended the basic unit test to ensure the name does not contain these
special characters and they are correctly removed.